refactor : [006-CUSTOMER_balance] 파일 구조 수정
Seller 파일로 SellerDto 이동
